[#60] Remove extend_knowledge_base
