+++ backend/app/api/api_v1/extended/extend_api.py
@@ -1,18 +1,12 @@
 from fastapi import APIRouter
 
 from app.api.api_v1.extended import (
-    extend_knowledge_base,
     prompt,
     system_settings,
 )
 
 api_router = APIRouter()
 
-api_router.include_router(
-    extend_knowledge_base.router,
-    prefix="/knowledge-base",
-    tags=["knowledge-base"],
-)
 api_router.include_router(prompt.router, prefix="/prompt", tags=["prompt"])
 api_router.include_router(
     system_settings.router,
